add the product_details and restuarent_view
